requirements.txt
sanic==22.6.1
python-dotenv==0.20.0
pytest==7.1.2
tortoise-orm==0.19.2
aerich==0.6.3
asyncpg==0.26.0 

geopy==2.2.0
plotly==5.9.0 
pandas==1.4.3
numpy==1.23.1
httpx==0.23.0 
kaleido==0.2.1
pytest-asyncio==0.19.0
sanic-testing==22.6.0  


pytest==7.1.2
black==22.6.0
flake8==5.0.4 